refactor(init): change tests
